split response component into 2 child components, success response and error response
