--- src/app/shared/type-check.pipe.ts
@@ -1,8 +1,0 @@
-import { Pipe, PipeTransform } from '@angular/core';
-
-@Pipe({
-  name: 'typeCheck'
-})
-export class ResponseTypeCheckPipe implements PipeTransform {
-  transform(value: any, ...propertyName: any): boolean {}
-}
