+++ src/app/tell-us/error-response/error-response.component.html
@@ -1,6 +1,5 @@
 <p class="alert warning">
-  We apologize, but there was an error with the form submission, please try
-  again later.
+  There was an error with the form submission, please try again later.
 </p>
 <dl>
   <dt>Error Code:</dt>
