+++ src/app/tell-us/error-response/error-response.component.ts
@@ -2,6 +2,9 @@ import { Component, Input } from '@angular/core';
 
 import { FeltReportReponseError } from './../felt-report-reponse-error';
 
+/**
+ * Error response component, shows error response after bad form submit
+ */
 @Component({
   selector: 'error-response',
   styleUrls: ['./error-response.component.scss'],
