--- src/app/tell-us/success-view/success-view.component.html
@@ -1,48 +1,0 @@
-<p class="alert success">
-  Thank you for your contribution. Your information will be processed shortly.
-</p>
-
-<dl>
-  <dt>Event Id:</dt>
-  <dd>{{ success.eventid }}</dd>
-
-  <dt>Estimated intensity:</dt>
-  <dd>
-    <shared-mmi
-      [bubble]="true"
-      [intensity]="success?.your_cdi | sharedRomanToNumber"
-      [value]="success?.your_cdi"
-    ></shared-mmi>
-  </dd>
-
-  <dt>Name:</dt>
-  <dd *ngIf="success.fldContact_name as name; else dash">{{ name }}</dd>
-
-  <dt>Email:</dt>
-  <dd *ngIf="success.fldContact_email as email; else dash">{{ email }}</dd>
-
-  <dt>Event time:</dt>
-  <dd>{{ success.ciim_time }}</dd>
-
-  <dt>Latitude:</dt>
-  <dd>{{ success.ciim_mapLat }}</dd>
-
-  <dt>Longitude:</dt>
-  <dd>{{ success.ciim_mapLon }}</dd>
-
-  <dt>Form version:</dt>
-  <dd>{{ success.form_version }}</dd>
-</dl>
-
-<div class="share-box">
-  <button
-    [disabled]="sdkStatus === false"
-    (click)="onSocialClick($event)"
-    class="fb-share-button"
-  >
-    <img src="./assets/facebook_share_button.jpg" alt="Share" />
-  </button>
-  <p>Share your DYFI response on Facebook.</p>
-</div>
-
-<ng-template #dash> <dd>&ndash;</dd> </ng-template>
